integerise

Generate integer cases from numeric weights matrix.

integerise(weights, inds, method = "trs", seed = 42)

Arguments

weights
A matrix or data frame of fractional weights, typically provided by rakeR::weight()

inds
The individual--level data (i.e. one row per individual)

method
The integerisation method specified as a character string. Defaults to "trs"; currently other methods are not implemented.

seed
The seed to use, defaults to 42.

Value

A data frame of integerised cases

Details

Extracted weights (using rakeR::extract()) are more 'precise' than integerised weights (although the user should be careful this is not spurious precision based on context) as they return fractions. Nevertheless, integerised weights are useful in cases when:

Numeric information (such as income) is required, as this needs to be cut() to work with rakeR::extract()

Simulated 'individuals' are required for case studies of key areas.

Input individual-level data for agent-based or dynamic models are required

The default integerisation method uses the 'truncate, replicate, sample' method developed by Robin Lovelace and Dimitris Ballas http://www.sciencedirect.com/science/article/pii/S0198971513000240

Other methods (for example proportional probabilities) may be implemented at a later date.

Examples

cons <- data.frame(
  "zone"   = letters[1:3],
  "a0_49"  = c(8, 2, 7),
  "a_gt50" = c(4, 8, 4),
  "f"      = c(6, 6, 8),
  "m"      = c(6, 4, 3),
  stringsAsFactors = FALSE
)

inds <- data.frame(
  "id"     = LETTERS[1:5],
  "age"    = c("a_gt50", "a_gt50", "a0_49", "a_gt50", "a0_49"),
  "sex"    = c("m", "m", "m", "f", "f"),
  "income" = c(2868, 2474, 2231, 3152, 2473),
  stringsAsFactors = FALSE
)
vars <- c("age", "sex")

weights     <- weight(cons = cons, inds = inds, vars = vars)
weights_int <- integerise(weights, inds = inds)
print(weights_int)

#>      id    age sex income zone
#> 1     A a_gt50   m   2868    a
#> 1.1   A a_gt50   m   2868    a
#> 2     B a_gt50   m   2474    a
#> 3     C  a0_49   m   2231    a
#> 3.1   C  a0_49   m   2231    a
#> 3.2   C  a0_49   m   2231    a
#> 3.3   C  a0_49   m   2231    a
#> 4     D a_gt50   f   3152    a
#> 4.1   D a_gt50   f   3152    a
#> 5     E  a0_49   f   2473    a
#> 5.1   E  a0_49   f   2473    a
#> 5.2   E  a0_49   f   2473    a
#> 5.3   E  a0_49   f   2473    a
#> 5.4   E  a0_49   f   2473    a
#> 1.2   A a_gt50   m   2868    b
#> 2.1   B a_gt50   m   2474    b
#> 2.2   B a_gt50   m   2474    b
#> 3.4   C  a0_49   m   2231    b
#> 4.2   D a_gt50   f   3152    b
#> 4.3   D a_gt50   f   3152    b
#> 4.4   D a_gt50   f   3152    b
#> 4.5   D a_gt50   f   3152    b
#> 5.5   E  a0_49   f   2473    b
#> 3.5   C  a0_49   m   2231    c
#> 3.6   C  a0_49   m   2231    c
#> 4.6   D a_gt50   f   3152    c
#> 4.7   D a_gt50   f   3152    c
#> 4.8   D a_gt50   f   3152    c
#> 5.6   E  a0_49   f   2473    c
#> 5.7   E  a0_49   f   2473    c
#> 5.8   E  a0_49   f   2473    c
#> 5.9   E  a0_49   f   2473    c
#> 5.10  E  a0_49   f   2473    c